Moved evaluate_dofs_* methods into a separate Python script
